Test and version images
The test will initialize, build and serve the demo site.  A number of
pages are retrieved as well.

To build the demo site successfully an upgrade to alpine:edge was needed.
A version conflict for setuptools dependencies prevents building the demo
site on alpine:3.4.

Images that pass the test will be tagged as the `latest` version as well
as the Nikola version (so people can upgrade at their own leisure).